--- src/main/java/com/hta2405/unite/action/BoardWriteAction.java
@@ -1,20 +1,0 @@
-package com.hta2405.unite.action;
-
-import java.io.IOException;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class BoardWriteAction implements Action {
-
-	@Override
-	public ActionForward execute(HttpServletRequest req, HttpServletResponse resp)
-			throws ServletException, IOException {
-		ActionForward forward = new ActionForward();
-		forward.setRedirect(false);
-		forward.setPath("/WEB-INF/views/board/boardWrite.jsp");
-		return forward;
-	}
-
-}
