--- src/main/java/com/hta2405/unite/action/EmpEmailVerificationAction.java
@@ -1,23 +1,0 @@
-package com.hta2405.unite.action;
-
-import java.io.IOException;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class EmpEmailVerificationAction implements Action {
-
-	@Override
-	public ActionForward execute(HttpServletRequest req, HttpServletResponse resp)
-			throws ServletException, IOException {
-		
-		
-		if(req.getSession().getAttribute("email")==null) {
-			req.getSession().setAttribute("message","세션이 만료되었습니다.");
-            return new ActionForward(true, "login");
-		}
-		return new ActionForward(false, "/WEB-INF/views/emp/emailVerificationForm.jsp");
-	}
-
-}
