--- src/main/java/com/hta2405/unite/action/EmpHomeAction.java
@@ -1,19 +1,0 @@
-package com.hta2405.unite.action;
-
-import java.io.IOException;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class EmpHomeAction implements Action {
-
-	@Override
-	public ActionForward execute(HttpServletRequest req, HttpServletResponse resp)
-			throws ServletException, IOException {
-		ActionForward forward = new ActionForward();
-		forward.setRedirect(false);
-		forward.setPath("/WEB-INF/views/emp/home.jsp");
-		return forward;
-	}
-}
