--- src/main/java/com/hta2405/unite/action/EmpLogoutAction.java
@@ -1,18 +1,0 @@
-package com.hta2405.unite.action;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-
-public class EmpLogoutAction implements Action {
-
-    @Override
-    public ActionForward execute(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        req.getSession().invalidate();//세션 초기화
-
-        return new ActionForward(true, req.getContextPath()+"/emp/login");
-    }
-}
