--- src/main/java/com/hta2405/unite/action/EmpPwInquiryAction.java
@@ -1,18 +1,0 @@
-package com.hta2405.unite.action;
-
-import java.io.IOException;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class EmpPwInquiryAction implements Action {
-
-	@Override
-	public ActionForward execute(HttpServletRequest req, HttpServletResponse resp)
-			throws ServletException, IOException {
-
-		return new ActionForward(false,"/WEB-INF/views/emp/pwInquiryForm.jsp");
-	}
-
-}
