+++ src/main/java/com/hta2405/unite/action/emp/EmpLogoutAction.java
@@ -1,0 +1,20 @@
+package com.hta2405.unite.action.emp;
+
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+
+public class EmpLogoutAction implements Action {
+
+    @Override
+    public ActionForward execute(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        req.getSession().invalidate();//세션 초기화
+
+        return new ActionForward(true, req.getContextPath()+"/emp/login");
+    }
+}
