+++ src/main/java/com/hta2405/unite/action/emp/EmpViewOtherDeptAction.java
@@ -1,0 +1,26 @@
+package com.hta2405.unite.action.emp;
+
+import java.io.IOException;
+
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+public class EmpViewOtherDeptAction implements Action {
+
+	@Override
+	public ActionForward execute(HttpServletRequest req, HttpServletResponse resp)
+			throws ServletException, IOException {
+		
+		
+		ActionForward forward = new ActionForward();
+		forward.setRedirect(false);
+		forward.setPath("/WEB-INF/views/empInfo/otherDeptInfo.jsp");
+		
+		return forward;
+	}
+
+}
