+++ src/main/java/com/hta2405/unite/controller/EmpFrontController.java
@@ -1,6 +1,6 @@
 package com.hta2405.unite.controller;
 
-import com.hta2405.unite.action.*;
+import com.hta2405.unite.action.emp.*;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
