+++ src/main/java/com/hta2405/unite/controller/ScheduleFrontController.java
@@ -1,8 +1,8 @@
 package com.hta2405.unite.controller;
 
-import com.hta2405.unite.action.ScheduleAddProcessAction;
-import com.hta2405.unite.action.ScheduleCalenderAction;
-import com.hta2405.unite.action.ScheduleListAction;
+import com.hta2405.unite.action.schedule.ScheduleAddProcessAction;
+import com.hta2405.unite.action.schedule.ScheduleCalenderAction;
+import com.hta2405.unite.action.schedule.ScheduleListAction;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
